[test] Replace useFakeTimers
